another fix

+++ tests/library_checker_aizu_tests/dsu/dsu_weighted_aizu.test.cpp
@@ -21,7 +21,7 @@ int main() {
       assert(dsu.size(v) == dsu_w.size(v));
     } else {
       ll curr_diff = dsu_w.diff(u, v);
-      if (curr_diff == 1e18) cout << "?\n";
+      if (curr_diff == ll(1e18)) cout << "?\n";
       else cout << dsu_w.diff(u, v) << '\n';
       assert(dsu.size(u) == dsu_w.size(u));
       assert(dsu.size(v) == dsu_w.size(v));
